weather_emulator/pyproject.toml
[project]
name = "weather_emulator"
version = "0.1.0"
description = "A mock weather data generator for MQTT"
requires-python = ">=3.9"
dependencies = [
    "paho-mqtt==1.6.1",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"